bot/requirements.txt
discord.py
supabase
python-dotenv
asyncio
